amelioration du control de nom d'une activity et d'une categorie ajout de la pagination des icons dans les categories

+++ src/app/category/category.component.ts
@@ -6,11 +6,7 @@ import { DataStorageService } from '../data-storage/data-storage.service';
 import { Category } from '../category';
 import { AVAILABLE_ICONS } from '../global';
 
-import { AbstractControl } from '@angular/forms';
-
-let nameFound: boolean = false;
-
-let okToSave: boolean = false;
+import { PageEvent } from '@angular/material';
 
 @Component({
   selector: 'app-category',
@@ -30,21 +26,21 @@ export class CategoryComponent implements OnInit {
 
   private nameFormControl: FormControl;
 
+  public pageEvent: PageEvent;
+  public pageSize: number;
+  public pageSizeOptions: number[];
+
   constructor(private router: Router, private route: ActivatedRoute, private dataBase: DataStorageService) {
     this.category = new Category();
     this.oldCategory = new Category();
     this.icons = AVAILABLE_ICONS;
-    this.nameFormControl = new FormControl('', [
-      Validators.required,
-      this.existNameValidator
-    ]);
+    this.nameFormControl = new FormControl('', [Validators.required]);
+    this.pageSizeOptions = [20, 50, 100];
+    this.pageSize = this.pageSizeOptions[0];
   }
 
-  existNameValidator(control: AbstractControl) {
-    if(nameFound){
-      return { existName: nameFound };
-    }
-    return null;
+  setExistName(v: boolean){
+    this.nameFormControl.setErrors({ existName: v });
   }
 
   ngOnInit() {
@@ -61,6 +57,10 @@ export class CategoryComponent implements OnInit {
         })
       }
     });
+
+    this.pageEvent = new PageEvent;
+    this.pageEvent.pageIndex = 0;
+    this.pageEvent.length = this.icons.length;
   }
 
   ngOnDestroy() {
@@ -71,10 +71,21 @@ export class CategoryComponent implements OnInit {
     this.subscription.unsubscribe();
   }
 
+  paginationFrom(pageEvent) {
+    return ((pageEvent.pageIndex === 0) ? pageEvent.pageIndex : (pageEvent.pageIndex) * pageEvent.pageSize );
+  }
+
+  paginationTo(pageEvent) {
+    if(pageEvent.pageSize){
+      this.pageSize = pageEvent.pageSize;
+    }
+    return this.paginationFrom(pageEvent) + this.pageSize;
+  }
+
   saveCategory() {
-    if(okToSave){
+    if(this.nameFormControl.errors == null){
       this.category.name = (((this.category.name).toLowerCase()).replace(/[\s]{2,}/g," ")).trim();
-      if (this.id != this.category.name) {
+      if (this.id && this.oldCategory.name != this.category.name) {
         this.dataBase.deleteCategory(this.oldCategory);
       }
       this.dataBase.saveCategory(this.category);
@@ -83,8 +94,7 @@ export class CategoryComponent implements OnInit {
   }
 
   verifyNameCategory(){
-    nameFound=false;
-    okToSave=false;
+    this.setExistName(false);
 
     let catInput = (((this.category.name).toLowerCase()).replace(/[\s]{2,}/g," ")).trim();
     let catPram = "";
@@ -102,20 +112,16 @@ export class CategoryComponent implements OnInit {
     }
 
     if(newCat){
-      this.nameFormControl = new FormControl(this.category.name, [Validators.required, this.existNameValidator]);
-      okToSave=true;
+      this.nameFormControl.setValue(this.category.name);
     }
     else{
       this.dataBase.getCategory(catInput).subscribe(cat => {
         var categoryFound = null;
         if(catInput != ""){
-          this.nameFormControl = new FormControl(this.category.name, [Validators.required, this.existNameValidator]);
+          this.nameFormControl.setValue(this.category.name);
           categoryFound = (this.category.name).toLowerCase();
-          if(catInput == catPram){
-            okToSave = true;
-          }
-          else{
-            nameFound=true;
+          if(catInput != catPram){
+            this.setExistName(true);
           }
         }
       });
